pyproject.toml
[project]
name = "moex-agent"
version = "1.0.0"
description = "MOEX Trading Signal Generator with Margin Risk Controls"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "PyYAML>=6.0.1",
    "pydantic>=2.6.0",
    "scikit-learn>=1.4.0",
    "joblib>=1.3.2",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
